Raptor Animator Setup and Idle Trigger
Initial trigger for the Idle animation has been set up using the keystroke "i" for now. IdleTrigger created in the Animator Window for the Raptor FBX object. Everything happening in the scene called RaptorScene found in assets. In the Animator Window, the idle animation is found and the IdleTrigger can be found in the parameters tab.